Refactor to make code easier to read
