+++ service/application/src/main/java/com/solace/maas/ep/event/management/agent/command/CommandManager.java
@@ -11,6 +11,7 @@ import com.solace.maas.ep.event.management.agent.plugin.service.MessagingService
 import com.solace.maas.ep.event.management.agent.plugin.solace.processor.semp.SempClient;
 import com.solace.maas.ep.event.management.agent.plugin.solace.processor.semp.SolaceHttpSemp;
 import com.solace.maas.ep.event.management.agent.plugin.terraform.manager.TerraformManager;
+import com.solace.maas.ep.event.management.agent.plugin.terraform.manager.TerraformUtils;
 import com.solace.maas.ep.event.management.agent.publisher.CommandPublisher;
 import com.solace.maas.ep.event.management.agent.util.MdcTaskDecorator;
 import lombok.extern.slf4j.Slf4j;
@@ -25,8 +26,6 @@ import java.util.Map;
 import java.util.concurrent.CompletableFuture;
 
 import static com.solace.maas.ep.event.management.agent.constants.Command.COMMAND_CORRELATION_ID;
-import static com.solace.maas.ep.event.management.agent.plugin.terraform.manager.TerraformManager.LOG_LEVEL_ERROR;
-import static com.solace.maas.ep.event.management.agent.plugin.terraform.manager.TerraformManager.setCommandError;
 
 @Slf4j
 @Service
@@ -62,7 +61,7 @@ public class CommandManager {
                 .exceptionally(e -> {
                     log.error("Error running command", e);
                     Command firstCommand = request.getCommandBundles().get(0).getCommands().get(0);
-                    setCommandError(firstCommand, (Exception) e);
+                    TerraformUtils.setCommandError(firstCommand, (Exception) e);
                     sendResponse(request);
                     return null;
                 });
@@ -75,7 +74,7 @@ public class CommandManager {
         } catch (Exception e) {
             log.error("Error getting terraform variables", e);
             Command firstCommand = request.getCommandBundles().get(0).getCommands().get(0);
-            setCommandError(firstCommand, e);
+            TerraformUtils.setCommandError(firstCommand, e);
             sendResponse(request);
             return;
         }
@@ -94,14 +93,14 @@ public class CommandManager {
                                     .logs(List.of(
                                             Map.of("message", "unknown command type " + command.getCommandType(),
                                                     "errorType", "UnknownCommandType",
-                                                    "level", LOG_LEVEL_ERROR,
+                                                    "level", TerraformUtils.LOG_LEVEL_ERROR,
                                                     "timestamp", OffsetDateTime.now())))
                                     .build());
                             break;
                     }
                 } catch (Exception e) {
                     log.error("Error executing command", e);
-                    setCommandError(command, e);
+                    TerraformUtils.setCommandError(command, e);
                 }
             }
         }
